Created "Smiley Buttons" and "Dice Roller" programs
Created a couple of programs for the BBC Micro:bit, a "Smiley Buttons" program (which shows different buttons based on button presses) and "Dice Roller" (turns the Micro:bit into a virtual dice). 

I'm still very much in the beginnings of my journey with learning hardware and as such a lot of this has just been basic tutorials, but have attempted to put my own spin on each of the basic lessons.