Added a "Pet Hamster" program
Probably amongst the silliest and most pointless of the tutorial programs on the MakeCode website, but I guess its a fun way of introducing the fact that the logo itself is an input of sorts. Once again, I tried to add a little to it, although admittedly not much in this instance.