Made a "Compass" program
Made a very simple Compass program that turns the Micro:bit into a very simple compass. This is the first program I've made outside of using the MakeCode editor, so it proved a little difficult to get it up and running as the Mu and Thonny MicroPython editors didn't let me flash the program to the Micro:bit, but the online Micro:bit Python Editor website allowed me to.